utils initialization
